requirements.txt
streamlit==1.36.0
openpyxl==3.1.5
pandas==2.2.2
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.92.0
